Rocket Launcher
Added rocket launcher

+++ Assets/Scripts/Damage.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 
 public class Damage
 {
-    public float damage;
+    public int damage;
     public bool isCritical;
     public float stoppingPower;
 
-    public Damage(float damage, bool isCritical, float stoppingPower){
+    public Damage(int damage, bool isCritical, float stoppingPower){
         this.damage = damage; // base damage
         this.isCritical = isCritical; // checks if critical (double damage)
         this.stoppingPower = stoppingPower; // reduces the speed of enemy
